requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
beautifulsoup4>=4.12.0 
pytz